+++ changelog.md
@@ -1,7 +1,10 @@
 ### Revision History
+* 3.6.8
+  * Bug fix: Contains search now searches column values for reference axes.
+  * Bug fix: Contains search now does not return a result from cube name.
 * 3.6.7
-  * Easily remove a transform from a reference axis
-  * NCube serialized through json-io now brings along ApplicationID
+  * Enhancement: Easily remove a transform from a reference axis
+  * Enhancement: NCube serialized through json-io now brings along ApplicationID
 * 3.6.6
   * Enhancement: Added custom reader and writer for NCube when used with json-io.
 * 3.6.5